### 2023-08-18 #5769 @editorialbot
[PRE REVIEW]: The Kestrel software for simulations of morphodynamic Earth-surface flows
**Submitting author:** @jakelangham (<a href="http://orcid.org/0000-0001-9857-7016">Jake Langham</a>)
**Repository:** https://github.com/jakelangham/kestrel/
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
