### 2023-08-24 fa549e7
Equivalent content in Maximums.txt to Maximums.nc Removed MaxHeights, MaxSpeeds etc. Amended documentation. Additional tidy-up.
