+++ src/Output.f90
@@ -38,7 +38,7 @@
 module output_module
 
    use set_precision_module, only: wp
-   use utilities_module, only: CheckFileExists, CheckPath, Int2String, KahanAdd, pair
+   use utilities_module, only: CheckFileExists, PathTrail, Int2String, KahanAdd, pair
    use grid_module, only: GridCoords, GridType, TileList, TileType
    use closures_module, only : FlowSquaredSpeedSlopeAligned, GeometricCorrectionFactor
    use runsettings_module, only : RunSet
@@ -152,13 +152,13 @@ contains
 
       logical dir_exists
 
-      inquire (file=RunParams%OutDir%s, exist=dir_exists)
+      inquire (file=RunParams%out_path%s, exist=dir_exists)
       if (dir_exists) then
-         call WarningMessage("Directory "//RunParams%OutDir%s//" exists and will be overwritten.")
+         call WarningMessage("Directory "//RunParams%out_path%s//" exists and will be overwritten.")
       else
          ! workaround: it calls an extern program...
-         call WarningMessage("Directory "//RunParams%OutDir%s//" will be created.")
-         call execute_command_line('mkdir -p '//trim(adjustl(RunParams%OutDir%s)), wait=.true.)
+         call WarningMessage("Directory "//RunParams%out_path%s//" will be created.")
+         call execute_command_line('mkdir -p '//trim(adjustl(RunParams%out_path%s)), wait=.true.)
       end if
    end subroutine CreateOutDir
 
@@ -202,16 +202,16 @@ contains
 #if HAVE_NETCDF4
       type(varString) :: MaximumFilenameNetCDF
 #endif
-      
+
       if (RunParams%out_txt) then
-         MaximumFilenameTXT = RunParams%OutDir + RunParams%MaximumsFilename + '.txt'
+         MaximumFilenameTXT = RunParams%out_path + RunParams%MaximumsFilename + '.txt'
          call OutputMaximums_txt(RunParams, MaximumFilenameTXT%s, grid)
       end if
 
       if (RunParams%out_kml) then
-         MaxHeightFilenameKML = RunParams%OutDir + RunParams%MaxHeightFilename + '.kml'
-         MaxSpeedFilenameKML = RunParams%OutDir + RunParams%MaxSpeedFilename + '.kml'
-         InundationTimeFilenameKML = RunParams%OutDir + RunParams%InundationTimeFilename + '.kml'
+         MaxHeightFilenameKML = RunParams%out_path + RunParams%MaxHeightFilename + '.kml'
+         MaxSpeedFilenameKML = RunParams%out_path + RunParams%MaxSpeedFilename + '.kml'
+         InundationTimeFilenameKML = RunParams%out_path + RunParams%InundationTimeFilename + '.kml'
          call OutputMaxHeights_kml(RunParams, MaxHeightFilenameKML%s, grid)
          call OutputMaxSpeeds_kml(RunParams, MaxSpeedFilenameKML%s, grid)
          call OutputInundationTime_kml(RunParams, InundationTimeFilenameKML%s, grid)
@@ -219,7 +219,7 @@ contains
 
 #if HAVE_NETCDF4
       if (RunParams%out_nc) then
-         MaximumFilenameNetCDF = RunParams%OutDir + RunParams%MaximumsFilename + '.nc'
+         MaximumFilenameNetCDF = RunParams%out_path + RunParams%MaximumsFilename + '.nc'
          call OutputMaximums_NetCDF(RunParams, MaximumFilenameNetCDF%s, grid)
       end if
 #endif
[…]
