+++ src/Output.f90
@@ -259,17 +259,253 @@ contains
 
       write (101, fmt="(a)") "Domain:"
 
+      write (101, fmt="(a,2P G0)") "Lat = ", RunParams%Lat
+      write (101, fmt="(a,3P G0)") "Lon = ", RunParams%Lon
+
+      write (101, fmt="(a)") "Boundary conditions = " // RunParams%bcs%s
+      if (RunParams%bcs%s=='dirichlet') then
+         write (101, fmt="(a,2P,G0)") "Boundary Hn = ", RunParams%bcsHnval
+         write (101, fmt="(a,2P,G0)") "Boundary U = ", RunParams%bcsuval
+         write (101, fmt="(a,2P,G0)") "Boundary V = ", RunParams%bcsvval
+         write (101, fmt="(a,2P,G0)") "Boundary psi = ", RunParams%bcspsival
+      end if
+
+      write (101, fmt="(a,i0)") "nXtiles = ", RunParams%nXtiles
+      write (101, fmt="(a,i0)") "nYtiles = ", RunParams%nYtiles
+      write (101, fmt="(a,i0)") "nXpertile = ", RunParams%nXpertile
+      write (101, fmt="(a,i0)") "nYpertile = ", RunParams%nYpertile
+      write (101, fmt="(a,G0)") "Xtilesize = ", RunParams%Xtilesize
+      write (101, *)
+
+      if (RunParams%set_Caps) then
+         do J = 1, RunParams%nCaps
+            write (101, fmt="(a)") "Cap:"
+            write (101, fmt="(a,G0)") "capX = ", RunParams%CapSources(J)%x
+            write (101, fmt="(a,G0)") "capY = ", RunParams%CapSources(J)%y
+            write (101, fmt="(a,G0)") "capRadius = ", RunParams%CapSources(J)%Radius
+            write (101, fmt="(a,G0)") "capVolume = ", RunParams%CapSources(J)%Volume
+            write (101, fmt="(a,G0)") "capHeight = ", RunParams%CapSources(J)%Height
+            write (101, fmt="(a,G0)") "capConc = ", RunParams%CapSources(J)%psi
+            write (101, fmt="(a,G0)") "capU = ", RunParams%CapSources(J)%U
+            write (101, fmt="(a,G0)") "capV = ", RunParams%CapSources(J)%V
+            write (101, fmt="(a,a)") "capShape = ", RunParams%CapSources(J)%Shape
+         end do
+      end if
+      if (RunParams%set_Cubes) then
+         do J = 1, RunParams%nCubes
+            write (101, fmt="(a)") "Cube:"
+            write (101, fmt="(a,G0)") "cubeX = ", RunParams%CubeSources(J)%x
+            write (101, fmt="(a,G0)") "cubeY = ", RunParams%CubeSources(J)%y
+            write (101, fmt="(a,G0)") "cubeLength = ", RunParams%CubeSources(J)%Length
+            write (101, fmt="(a,G0)") "cubeWidth = ", RunParams%CubeSources(J)%Width
+            write (101, fmt="(a,G0)") "cubeVolume = ", RunParams%CubeSources(J)%Volume
+            write (101, fmt="(a,G0)") "cubeHeight = ", RunParams%CubeSources(J)%Height
+            write (101, fmt="(a,G0)") "cubeConc = ", RunParams%CubeSources(J)%psi
+            write (101, fmt="(a,G0)") "cubeU = ", RunParams%CubeSources(J)%U
+            write (101, fmt="(a,G0)") "cubeV = ", RunParams%CubeSources(J)%V
+            write (101, fmt="(a,a)") "cubeShape = ", RunParams%CubeSources(J)%shape
+         end do
+      end if
[…]
