### 2023-08-25 2a8f6c3
Equivalent content in Maximums.txt to Maximums.nc (#4)
* Equivalent content in Maximums.txt to Maximums.nc
Removed MaxHeights, MaxSpeeds etc.
Amended documentation.
Additional tidy-up.

* Small typo corrections to output.rst

* Moved extra data in RunInfo to end
Important change in DEM directory search

---------

Co-authored-by: Jake Langham <46025329+jakelangham@users.noreply.github.com>
