### 2023-08-25 #5769 @kthyng
@crvernon Are you up for editing this submission? Not sure how far into geology you are...
