+++ paper/paper.bib
@@ -31,7 +31,8 @@
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://demichie.github.io/IMEX_SfloW2D/}
+  url = {https://demichie.github.io/IMEX_SfloW2D/},
+  doi = {10.5194/gmd-2023-80}
 }
 
 @article{Vitturi:2023b,
@@ -64,7 +65,8 @@ DOI = {10.5194/egusphere-2023-1301}
   number={1-2},
   pages={1--21},
   year={2005},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jvolgeores.2004.06.014}
 }
 @inproceedings{Simakov:2019,
   title={Modernizing {Titan2D}, a parallel {AMR} geophysical flow code to support multiple rheologies and extendability},
@@ -72,7 +74,8 @@ DOI = {10.5194/egusphere-2023-1301}
   booktitle={Int. C. High Perform.},
   pages={101--112},
   year={2019},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-030-34356-9_10}
 }
 @article{Patra:2020,
   title={Comparative Analysis of the Structures and Outcomes of Geophysical Flow Models and Modeling Assumptions Using Uncertainty Quantification},
@@ -105,7 +108,8 @@ DOI = {10.5194/egusphere-2023-1301}
   number={2170},
   pages={20130820},
   year={2014},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+  doi={10.1098/rspa.2013.0820}
 }
 
 @article{Pudasaini:2019,
@@ -127,7 +131,8 @@ DOI = {10.5194/egusphere-2023-1301}
   number={2},
   pages={265--276},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10346-021-01733-2}
 }
 
 @article{Christen:2010,
@@ -138,7 +143,8 @@ DOI = {10.5194/egusphere-2023-1301}
   number={1-2},
   pages={1--14},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.coldregions.2010.04.005}
 }
 
 @article{Jenkins:2023,
@@ -148,7 +154,8 @@ DOI = {10.5194/egusphere-2023-1301}
   volume={84},
   pages={103338},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ijdrr.2022.103338}
 }
 @misc{Langham:2023,
       title={Simulating shallow morphodynamic flows on evolving topographies}, 
@@ -183,7 +190,8 @@ DOI = {10.5194/egusphere-2023-1301}
   journal={J. Fluid Mech.},
   volume={916},
   year={2021},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/jfm.2021.235}
 }
 
 @article{Capart:1998,
@@ -193,7 +201,8 @@ DOI = {10.5194/egusphere-2023-1301}
   volume={372},
   pages={165--187},
   year={1998},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/s0022112098002250}
 }
 
 @book{Jakob:2005,
