### 2023-08-30 ea5293e
a few more DOIs
