+++ paper/paper.bib
@@ -31,8 +31,7 @@
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
-  url = {https://demichie.github.io/IMEX_SfloW2D/},
-  doi = {10.5194/gmd-2023-80}
+  url = {https://demichie.github.io/IMEX_SfloW2D/}
 }
 
 @article{Vitturi:2023b,
@@ -42,7 +41,8 @@
   volume={2023},
   pages={1--42},
   year={2023},
-  publisher={G{\"o}ttingen, Germany}
+  publisher={G{\"o}ttingen, Germany},
+  doi = {10.5194/gmd-2023-80}
 }
 
 @Article{Vitturi:2023a,
@@ -120,7 +120,8 @@ DOI = {10.5194/egusphere-2023-1301}
   number={12},
   pages={2920--2942},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.1029/2019JF005204}
 }
 
 @article{Meyrat:2022,
@@ -210,7 +211,8 @@ DOI = {10.5194/egusphere-2023-1301}
   author={Jakob, M. and Hungr, O.},
   volume={739},
   year={2005},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/b138657}
 }
 
 @Article{Iverson:1997,
