### 2023-08-30 #5769 @jakelangham
@crvernon Hi - in preparation for review I have added the remaining DOIs as suggested by the bot.

Please note: the preprint currently has a broken reference on line 41 that I don't know how to fix (I assume it's a pandoc issue since the .bib file looks ok).
