### 2023-11-13 #5769 @dlgeorge
Hi Chris,

thanks for the invite, looks interesting, but I'm overcommitted at the
moment.

verty best,

-dave

On Fri, Nov 3, 2023 at 7:44 AM Chris Vernon ***@***.***>
wrote:

> 👋 @dlgeorge <https://github.com/dlgeorge> - Would you be willing to
> review this submission to JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
