### 2023-11-21 #6079 @editorialbot
[REVIEW]: The Kestrel software for simulations of morphodynamic Earth-surface flows
**Submitting author:** @jakelangham (<a href="http://orcid.org/0000-0001-9857-7016">Jake Langham</a>)
**Repository:** https://github.com/jakelangham/kestrel/
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
