### 2023-11-22 18fa79e
Show how to specifiy GDAL/PROJ dirs in quickstart
