### 2023-11-22 06d96d3
Mention requirement of NetCDF-Fortran for nc support
