### 2023-11-22 7f04540
Emphasise auto{conf,make} requirement in README.md
