### 2023-11-22 0db5261
Check failure of output during netcdf restart test
