### 2023-12-01 jakelangham/kestrel#11 @mdpiper
[joss review] Minor spelling errors in documentation
I came across a couple spelling errors in the docs (e.g., on lines 28 and 40 of `quickstart.rst`). Running `ispell` over the docs should fix this.

See https://github.com/openjournals/joss-reviews/issues/6079.
