### 2023-12-04 jakelangham/kestrel#12 @jatkinson1000
Test script failure on mac
The test script fails on mac due to there not being an `ldd` command ([line 50, `tests/runall.jl`](https://github.com/jakelangham/kestrel/blob/022fe65f9ff0b1852f6875fcffecbf3954b751f7/tests/runall.jl#L50)) on mac OS.

This can be resolved by using `otool -L` in place of `ldd`, and indeed, if I do this the tests run fine.

Consider making this point in the documentation, or implementing both in julia (OS detection, or try and except with `ldd`)?

Discovered as part of https://github.com/openjournals/joss-reviews/issues/6079
