### 2023-12-04 #6079 @jatkinson1000
@jakelangham a couple of points as I work through this:

- I had to do some work to run the tests and have opened issue https://github.com/jakelangham/kestrel/issues/12
- The web docs have no information about running the tests (only the README.md).
   Could you repeat the testing info on the web docs please?
- Similarly, the contribution guidelines are only available on the web documentation.
   Consider adding a short section (even if just a link) in the GitHub README.md
- Whilst there _is_ a statement of need it is very brief.
   Consider expanding it slightly with more description. This should also help in making the software discoverable.
   Consider also adding tags and a website link to the GitHub 'about'.
- There is no API or subroutine documentation as far as I can tell.
   I suppose the argument is that this is designed to be used as complete code with varying input files (which are very well documented)?
   If you want contributions/development consider more detailed API documentation using something such as [FORD](https://forddocs.readthedocs.io/en/latest/), though I understand this may be be beyond the scope of this review.

I'll now read the paper in more detail and follow up with related comments.
