### 2023-12-04 #6079 @jakelangham
@jatkinson1000 I'll work through these in due course.

I actually didn't realise a statement of need was required for the docs, but it makes complete sense.

Regarding API documentation, it would be good to clarify this before proceeding much further and to get your opinion. We thought about this issue before submission and came to the view that an independent auto-generated API documentation was a bit too heavy duty for our purposes. While it is a sizeable codebase, the control flow is pretty linear and the routines that are 'library-like' in their design are probably in the minority. This is my own prejudice, but I have seen a fair few research codes where the auto-generated API docs were totally unhelpful compared with just reading the source code.

I remember reading somewhere in the JOSS docs that for Fortran/C++ codes, having separate API documentation was optional (but encouraged), provided that the source code was well commented. So at the time we made the decision to concentrate on improving our source, since that seemed like a more obvious benefit to us, rather than spending time on a more formal API. We were aiming to make sure every module had a explanatory header + each of the most important routines, so hopefully you should find that this is the case. Let me know whether you think this is sufficient.
