### 2023-12-04 #6079 @jatkinson1000
@jakelangham with regards to the paper:

- Not super important, but it is not clear to me why the project is called `kestrel`.
   It does not appear to be an acronym. Explaining this would be nice, and I feel somewhat useful.
- Line 41 there is an issue with reference `Iverson2015`
- I don't understand the date range in reference `Mergili (2014-2023)`. Is this a typo/bibtex error?
- There are several alternative codes listed, including a number which are open source.
   The principal argument for kestrel seems to be that 'others are not well documented'. I would really like to see justification beyond this - why did you choose to create a new model rather than contributing to improve the documentation of an existing open-source model?
   This is made clearer from line 91, but I suggest restructuring to make it clear when talking about other codes that kestrel has different physics (Langham et al. (2023)) to improve morphodynamic modelling and ease of use. This is, in my view, the main motivation.
   As far as I can tell this is companion software to a new modelling approach described in Langham et al. (2023), and hence not available in the other codes.
   Being better documented than others is, to me, a corollary of this being quality research software, not a primary motivation.
- The figure [here](https://kestrel-unibristol.readthedocs.io/en/latest/physical%20model.html) provides a really nice illustration of the concepts discussed in the theoretical background. Consider adding it to the paper.
- The final sentence of the penultimate paragraph (starting line 98 "These include...") doesn't read right to me and seems out of place. What is "These" referring to?
- When describing the 'many free parameters' perhaps refer to the online input documentation which has longer description of these?
- The paper can, at times, sound a little defensive of the fact that kestrel is more simple than other codes.
   Perhaps include a brief point about the importance of reduced order modelling for physical understanding (you do hint at this to a degree). I have been in a similar situation with codes I have written so am familiar.

Overall I think this is well-written and organised software, and it is clearly a significant piece of work. I am more than happy to recommend publication in JOSS subject to resolving the minor issues raised above, the points which I believe will strengthen the paper, and any comments from @mdpiper.
