### 2023-12-05 4a3799b
Add basic mac support to netcdf test
