### 2023-12-05 b7f037a
Add basic help dialogue to test suite
