### 2023-12-05 f80a288
Add info about running tests to quickstart
