### 2023-12-05 7a7c83c
Add contribution info to README.md
