### 2023-12-05 #6079 @jakelangham
@jatkinson1000 A few updates...

> @jakelangham a couple of points as I work through this:
>
>     * [ ]  I had to do some work to run the tests and have opened issue [Test script failure on mac jakelangham/kestrel#12](https://github.com/jakelangham/kestrel/issues/12)

https://github.com/jakelangham/kestrel/commit/4a3799b9f39d25e88d01aebeb3dbb44fae982d0a fixes this I believe.

>     * [ ]  The web docs have no information about running the tests (only the README.md).
>       Could you repeat the testing info on the web docs please?

Done, in https://github.com/jakelangham/kestrel/commit/f80a28837dc4c4428909529e25944778876851bf

>     * [ ]  Similarly, the contribution guidelines are only available on the web documentation.
>       Consider adding a short section (even if just a link) in the GitHub README.md

Done, in https://github.com/jakelangham/kestrel/commit/7a7c83ccc9b160c9cfc914b3a067e06990b71e8f
