### 2023-12-05 cba2c7f
Convert string with multiple command line args to list for apple NetCDF check.
