### 2023-12-05 jakelangham/kestrel#12 @jatkinson1000
@jakelangham Mostly good, just a small fix that I have submitted in PR #13
