### 2023-12-05 d1d99a0
Fix julia test script to run on mac (#13)
* Convert string with multiple command line args to list for apple NetCDF check.

* Reimpose whitespace formatting as original.
