### 2023-12-06 81b39cc
Add features list to README.md
