+++ paper/paper.bib
@@ -22,7 +22,7 @@
   title = {r.avaflow -- The mass flow simulation tool},
   author = {Mergili, M., Pudasaini, S.P.},
   url = {https://www.avaflow.org},
-  year = {2014-2023},
+  year = {2023},
 }
 
 @misc{IMEX_SfloW2D:2023,
