### 2023-12-06 79c19c3
Bullet pts for eddy visc & bcs
