### 2023-12-06 f506363
Add refs to README and clarify well-posedness
