+++ README.md
@@ -12,8 +12,7 @@ closures to be selected at runtime to specialise simulations for particular purp
 
 * Fully conservative well-balanced positivity-preserving finite volume solver.
 * Dynamic erosion and deposition of sediment.
-* Eddy viscosity implementation to ensure well-posedness of morphodynamic flows
-  (see [here](https://arxiv.org/abs/2007.15989) and [here](https://arxiv.org/abs/2306.16185) for detailed discussions of this issue).
+* Eddy viscosity implementation to ensure well-posedness of morphodynamic flows.
 * User-settable basal drag and morphodynamics parametrisations.
 * Simulation on user-specified topographies: simple geometric surfaces and digital elevation maps (DEMs).
 * Variety of boundary conditions for initiating flows, including from multiple locations.
@@ -22,7 +21,9 @@ closures to be selected at runtime to specialise simulations for particular purp
 An overview of the physical modelling framework may be found 
 [here](https://kestrel-unibristol.readthedocs.io/en/latest/physical%20model.html).
 For full details, including the numerical implementation, consult our 
-[paper](https://arxiv.org/abs/2306.16185).
+paper [[1]](https://arxiv.org/abs/2306.16185). The issue of well-posedness referred to above 
+is explored in detail in refs [[1]](https://arxiv.org/abs/2007.15989) and 
+[[2]](https://arxiv.org/abs/2306.16185).
 
 ## 'Very' quick start ##
 
@@ -96,3 +97,9 @@ contributions that might be valuable is given [here](https://kestrel-unibristol.
 Kestrel is developed and maintained by Mark J. Woodhouse
 (mark.woodhouse@bristol.ac.uk) and Jake Langham (J.Langham@bristol.ac.uk),
 University of Bristol.
+
+## References
+
+* [1] Langham J, Woodhouse MJ, Hogg AJ, Jenkins LT, Phillips JC. 2023 Simulating shallow morphodynamic flows on evolving topographies. _arXiv_ **2306.16185**
+* [2] Langham J, Woodhouse MJ, Hogg AJ, Phillips JC. 2021 Linear stability of shallow
+morphodynamic flows. _J. Fluid Mech._ **916**.
