### 2023-12-06 acac18e
Add features list to README.md (#14)
* Add features list to README.md

* Bullet pts for eddy visc & bcs

* Add refs to README and clarify well-posedness issue
