### 2023-12-06 #6079 @jakelangham
>     * [ ]  Whilst there _is_ a statement of need it is very brief.
>       Consider expanding it slightly with more description. This should also help in making the software discoverable.
>       Consider also adding tags and a website link to the GitHub 'about'.

@jatkinson1000 We have updated the github README.md now and added some tags to address this suggestion.
