### 2023-12-06 788323d
Minor corrections/adjustments to paper
