+++ paper/paper.bib
@@ -11,7 +11,7 @@
 }
 @misc{DCLAW:2023,
   title = {{D-Claw}: software for granular-fluid flows},
-  author = {George, D. and Barnhart, K.},
+  author = {George, D. L. and Barnhart, K.},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
