+++ paper/paper.md
@@ -99,7 +99,11 @@ uses a slightly different description of the flow physics and underlying
 mathematical framework. In some cases, it can be difficult to discern from the
 available documentation exactly which model and assumptions are used in the
 latest software version and how the program operates `under the hood' to produce
-its results. 
+its results.  The diversity of approaches reflects differences in the level of
+detail included in physical descriptions, as well as genuine uncertainties
+present in current understanding of Earth surface flow physics.  Our code implements
+a newly derived modelling framework and accompanying numerical scheme detailed
+in @Langham:2023. Its essential features are described below.
 
 Like all the above software, `Kestrel` numerically approximates solutions to an
 underlying system of partial differential equations for the flow, whose
@@ -129,23 +133,21 @@ $\theta(\mathbf{x},t)$ is the local slope angle between the bed normal and
 gravity, and $\nabla_s = \nabla - \mathbf{s}(\mathbf{s}\cdot\nabla)$, 
 with $\mathbf{s} \equiv \cos(\theta)\nabla b$.
 
-The technical details of these equations, their derivation and our numerical
-solution scheme are fully presented by @Langham:2023.  While most of the terms
-are fixed by the underlying depth-averaged flow physics (and shall not be
-discussed further), some parts of the right-hand sides are user-settable.  The
-terms $\mathcal{T}$, $\mathcal{E}$ and $\mathcal{D}$ denote the basal drag,
-erosion rate and deposition rate respectively. These are modelling closures,
-assumed to be functions of the flow fields $H$, $\bar{\mathbf{u}}$ and
-$\bar{\psi}$. In each case, the user may choose from different options,
-depending on the problem at hand. For example, the drag $\mathcal{T}$ may be
-set either to a function appropriate for turbulent fluids, to various models of
-purely granular flows, or to a combined law that depends on the solids
-concentration.  This provides the flexibility to simulate many different kinds
-of flow.  Furthermore, it is worth noting that in many cases, the question of
-which closures most faithfully capture the flow physics is an open problem that
-cannot easily be addressed experimentally.  Using numerical simulations to
-investigate the effects of different modelling choices is one way to approach
-this.
+While most of the terms are fixed by the underlying depth-averaged flow physics
+(and shall not be discussed further), some parts of the right-hand sides are
+user-settable.  The terms $\mathcal{T}$, $\mathcal{E}$ and $\mathcal{D}$ denote
+the basal drag, erosion rate and deposition rate respectively. These are
+modelling closures, assumed to be functions of the flow fields $H$,
+$\bar{\mathbf{u}}$ and $\bar{\psi}$. In each case, the user may choose from
+different options, depending on the problem at hand. For example, the drag
+$\mathcal{T}$ may be set either to a function appropriate for turbulent fluids,
+to various models of purely granular flows, or to a combined law that depends on
+the solids concentration.  This provides the flexibility to simulate many
+different kinds of flow.  Furthermore, it is worth noting that in many cases,
+the question of which closures most faithfully capture the flow physics is an
+open problem that cannot easily be addressed experimentally.  Using numerical
+simulations to investigate the effects of different modelling choices is one way
+to approach this.
 
 The remaining source terms $\mathcal{Q}_H$ and $\mathcal{Q}_{\psi}$ are
 time-dependent functions that provide one way for a modeller to control fluxes
