### 2023-12-06 7ec1577
Add note that parameters/closures listed in docs
