### 2023-12-06 #6079 @jatkinson1000
Thanks @jakelangham, that's looking good.

>https://github.com/jakelangham/kestrel/commit/ee063490212a986c79639e39edb3b2cd381d7182 removes the 2014- part. This is one of several website references I included when citing alternative codes and to be honest, I don't know what the right convention is for the year of publication. Listing them all as 2023 doesn't feel quite right when some codes have been around for a decade or so.

Right, I follow. My understanding is that you cite the 'date accessed' for web resources. Ideally there would be a JOSS/GMD paper or repository with a DOI you could also cite, but these are not common for older (and indeed many current :cry:) codes. Here I would go with 2014 if it hasn't changed since then, or 2023 if it has. Looking at avaflow it has release numbers, so I would add that to the citation with date 2023 to be clear what you are comparing this work to.

>The figure itself would need to be modified a little to align it with the paper. I'm not against this in principle, assuming it doesn't conflict with any length requirements, but would prefer to wait for @mdpiper's review before doing this.

I personally feel it would greatly improve the readability of the paper, but happy to wait for @mdpiper 's opinion.

>I've tried to address this in https://github.com/jakelangham/kestrel/commit/61aa2b3cac2fe59a2c2d83e89c365767c7dcc887. My inclination was to preserve the 'line 91' paragraph as is, since it follows the description of the equations, but highlight up front earlier that the model framework and scheme are new.

I agree with not being too technical early on, but my issue at the moment is that when reading the paper it comes across that the principal reason for developing `kestrel` is better documentation and an inability to understand other models.
In reality I think the case is that kestrel models different physics to the alternatives, and this is why a new code has been written.
I believe it would greatly help the paper to make this clear when introducing the other models - you can leave the technical discussion of the physics until later, but a layperson's description something like "kestrel is capable of simulating a wide range of flow types from fluid to granular, not possible with a single alternative code, and implements the new modelling approach defined in Langham (2023) and expanded upon below" (this may not be quite correct as I am not overly familiar, but you get the idea).
I would also put this ahead of the comments on documentation as I think it is a more important motivation.
