### 2023-12-07 #6079 @jakelangham
@jatkinson1000 Ok, I take the point - it should be possible to rewrite closer to what you're suggesting, with the caveat that its difficult to compare all these codes directly. The phrase about documentation could even be removed to be honest. Like the figure issue, I think I'll park this for now and come to it with a fresh perspective when addressing the next round of review comments.
