### 2023-12-07 26284bb
Simpler fix to NetCDF
