### 2023-12-08 5ee8bee
Tidy up unused variables
