### 2023-12-08 2bde668
Netcdf (#10)
* Fixes to Restart, include basePath and georeferencing

* Fix to NetCDF

* Removed unrelated flume code

Reason: not related to fix

* Simpler fix to NetCDF

* Tidy up unused variables

---------

Co-authored-by: Jake Langham <46025329+jakelangham@users.noreply.github.com>
Co-authored-by: Jake Langham <jakelangham@mykolab.com>
