### 2023-12-08 fa1641d
RebuildDEM option now working.  Additional fix of geotif handling for some coordinate systems.
