### 2023-12-09 e8569a8
Rebuilt dem (#15)
* RebuildDEM option now working.  Additional fix of geotif handling for some coordinate systems.

---------

Co-authored-by: Jake Langham <46025329+jakelangham@users.noreply.github.com>
Co-authored-by: Jake Langham <jakelangham@mykolab.com>
