### 2023-12-18 299d518
Add schematic to paper
