### 2023-12-19 64585b3
Add versions to software refs where applicable
