+++ paper/paper.bib
@@ -1,15 +1,15 @@
-@misc{Laharflow:2023,
+@software{Laharflow:2023,
   title = {LaharFlow: the {B}ristol model of lahars on evolving topography},
   author = {Woodhouse, M. J.},
   year = {2023},
   url = {https://laharflow.bris.ac.uk/}
 }
-@misc{Flo2D:2023,
+@software{Flo2D:2023,
   title = {Flo-2D},
   year = {2023},
   url = {https://flo-2d.com/}
 }
-@misc{DCLAW:2023,
+@software{DCLAW:2023,
   title = {{D-Claw}: software for granular-fluid flows},
   author = {George, D. L. and Barnhart, K.},
   year = {2023},
@@ -18,19 +18,21 @@
   url = {https://github.com/geoflows/D-Claw}
 }
 
-@misc{ravaflow:2023,
+@software{ravaflow:2023,
   title = {r.avaflow -- The mass flow simulation tool},
-  author = {Mergili, M., Pudasaini, S.P.},
+  author = {Mergili, M. and Pudasaini, S. P.},
   url = {https://www.avaflow.org},
+  version = {3},
   year = {2023},
 }
 
-@misc{IMEX_SfloW2D:2023,
+@software{IMEX_SfloW2D:2023,
   title = {IMEX_SfloW2D},
   author = {{de' Michieli Vitturi}, M. and Lari, G.},
   year = {2023},
   publisher = {GitHub},
   journal = {GitHub repository},
+  version = {1.0.0},
   url = {https://demichie.github.io/IMEX_SfloW2D/}
 }
 
