### 2023-12-19 #6079 @jakelangham
@jatkinson1000

I've made a few improvements which I think address the remainder of your unresolved comments. Specifically, https://github.com/jakelangham/kestrel/commit/299d518c6e5bbe98a1c89315583254c5b747e3e5 adds a slightly revised version of the flow diagram from the docs, https://github.com/jakelangham/kestrel/commit/3c66587314e9de0f4127de1b4cc03a8981105911 rejigs the statement of need to clarify our contribution a bit better and https://github.com/jakelangham/kestrel/commit/64585b34e81184dff76e72f15662115bac286379 adds version numbers to software references where I could find them.
