### 2023-12-19 #6079 @jatkinson1000
Hi @crvernon @jakelangham @mdpiper

I'm happy with the paper and software now - both are well-written and I recommend publication in the latest state, hopefully in time for the end of the year! 🚀

Considering @mdpiper's comments I don't see anything to prevent publication, but encourage packaging in future (fpm is a good suggestion), and agree that the option of setting the executable location somewhere other than `src/` would be a nice additional feature. Whilst wide compiler support is nice, depending on GCC is at least much less restrictive than depending on the intel suite etc.
