### 2023-12-19 d55bab5
Note about ./configure --prefix in docs
