### 2023-12-19 #6079 @jatkinson1000
Nothing special on mac, but yes, I used gcc installed via homebrew (same for netcdf and other dependencies).

RE: Fortran vs C++ as a numeric code there could perhaps be an argument made based on Fortran's native array handling/capabilities?
