### 2023-12-21 1ef5201
Add note about finding/installing dependencies
