### 2023-12-21 #6079 @jakelangham
Ok thanks, I have added a comment about homebrew to the dependencies section of the docs https://github.com/jakelangham/kestrel/commit/1ef52018f331c4c71d34385279f3eb22a1092d51
