### 2024-01-14 #6079 @crvernon
@jakelangham I am having some trouble compiling the paper into the JATS (XML format) version of the paper.  This is due to the compilation for JATS using pandoc instead of the PDF build that happens via LaTeX.

I asked another editor about this and they recommend splitting up the single `gather` environment with four equations in it into four separate `equation` environments.  See this issue:  https://github.com/lierdakil/pandoc-crossref/issues/131

Could you give this a try?  There is no need to conduct another release, etc.  You only need to make changes to the paper that is on `main` and I can give it another shot.

Thanks!
