### 2024-01-15 fda9f1a
Convert gather to 4 equation envs for xml output
