### 2024-01-15 #6079 @markwoodhouse
@jakelangham I've just read through the proof, and all looks good to me.
