+++ paper/paper.bib
@@ -1,3 +1,14 @@
+@article{Phillips:2024,
+  title={Impacts and prospective hazard analysis of rainfall-triggered lahars on {S}t {V}incent 2021--22},
+  author={Phillips, J. and Barclay, J. and Cole, P. and Johnson, M. and Miller, V. and Robertson, R.},
+  journal={Geol. Soc., Lond., Spec. Publ.},
+  volume={539},
+  number={1},
+  pages={245--266},
+  year={2024},
+  publisher={The Geological Society of London},
+  doi={10.1144/SP539-2022-313}
+}
 @software{Laharflow:2023,
   title = {LaharFlow: the {B}ristol model of lahars on evolving topography},
   author = {Woodhouse, M. J.},
