### 2024-01-20 f0b9310
Update paper.md (#17)
* Update paper.md

* Add new reference

---------

Co-authored-by: Jake Langham <jakelangham@mykolab.com>
