### 2024-01-20 #6079 @jakelangham
@kthyng Thanks, please note that I have taken the liberty of modifying your pull request to include an additional reference that was published recently, see https://github.com/jakelangham/kestrel/commit/f0b93102520f4ae99c4f8aab1e2dcdc1ec907aff
