### 2024-01-22 b136192
Output as .txt if Format field not recognised
