### 2024-01-22 c9ba568
Always store erosion choice in .nc files
